pyproject.toml
[tool.poetry]
name = "sdm-ml-bc"
version = "0.1.0"
description = "Machine learning development for neoadjuvant chemotherapy reponder of bladder cancer"
authors = ["Jinahn J <jinnannann@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.1"
pandas = "^2.2.3"
pillow = "^10.4.0"
imageio = "^2.35.1"
scikit-learn = "^1.5.2"
scipy = "^1.14.1"
matplotlib = "^3.9.2"
imblearn = "^0.0"
graphviz = "^0.20.3"
ipython = "^8.27.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
